Moved error notice functionality into its own class and changed to all use the same variable to hide things
