+++ classes/WpMatomo.php
@@ -38,7 +38,6 @@ use WpMatomo\User\Sync as UserSync;
 
 class WpMatomo {
 
-	const OPTION_NAME_SYSTEM_REPORT_ERRORS_DISMISSED = 'matomo_system_report_errors_dismissed';
 	/**
 	 * @var Settings
 	 */
@@ -99,7 +98,8 @@ class WpMatomo {
 				$referral->register_hooks();
 			}
 
-			add_action( 'admin_notices', [ $this, 'check_errors' ] );
+			$errorNotice = new \WpMatomo\ErrorNotice( self::$settings );
+			$errorNotice->register_hooks();
 
 			$chart = new Chart();
 			$chart->register_hooks();
@@ -127,31 +127,6 @@ class WpMatomo {
 				'add_settings_link',
 			]
 		);
-	}
-
-	public function check_errors() {
-		if ( isset( $_GET['page'] ) && substr( sanitize_text_field( wp_unslash( $_GET['page'] ) ), 0, 7 ) === 'matomo-' ) {
-			$system_report = new \WpMatomo\Admin\SystemReport( self::$settings );
-			if ( $system_report->errors_present() ) {
-				add_action(
-					'admin_notices',
-					function () {
-						echo '<div class="notice notice-warning is-dismissible" id="matomo-systemreporterrors">
-                      <p>' . esc_html__( 'There are some errors in the', 'matomo' ) .
-							' <a href="' . esc_url( admin_url( 'admin.php?page=matomo-systemreport' ) ) . '">' . esc_html__( 'Matomo Diagnostics System report', 'matomo' ) . '</a> ' .
-							esc_html__( 'that may prevent the plugin for working normally.', 'matomo' ) . '</p></div>';
-					}
-				);
-				add_action(
-					'wp_ajax_matomo_system_report_error_dismissed',
-					function () {
-						if ( is_admin() ) {
-							update_user_meta( get_current_user_id(), self::OPTION_NAME_SYSTEM_REPORT_ERRORS_DISMISSED, true );
-						}
-					}
-				);
-			}
-		}
 	}
 
 	private function check_compatibility() {
